sqlccl: move backup/restore/load benchmarks to a common file
And delete one of the redundant Restore benchmarks.
